refactor and fix problems in cart cubit and meal details view and address view

+++ lib/features/cart/data/repository/cart_abstract_repo.dart
@@ -3,7 +3,7 @@ import 'package:restaurant/features/cart/data/models/cart_model.dart';
 abstract class CartAbstractRepository {
   Future<CartModel> getCart();
   Future<void> fetchCart();
-  Future<void> addToCart({required int dishId, required int price});
+  Future<void> addToCart({required int dishId, required double price});
   Future<void> updateCartItem({required int itemId, required int quantity});
 
   Future<void> deleteCartItem(int itemId);
